non-functional - moved validation utils file to parent folder, for future code use in new services/comments folder.
