+++ legal-api/src/legal_api/services/filings/validations/change_of_name.py
@@ -20,7 +20,7 @@ from flask_babel import _
 from legal_api.errors import Error
 from legal_api.models import Business
 
-from ..utils import get_str
+from ...utils import get_str
 
 
 def validate(business: Business, con: Dict) -> Error:
