+++ legal-api/src/legal_api/services/filings/validations/voluntary_dissolution.py
@@ -20,7 +20,7 @@ from flask_babel import _
 from legal_api.errors import Error
 from legal_api.models import Business
 
-from ..utils import get_bool, get_date
+from ...utils import get_bool, get_date
 
 
 def validate(business: Business, con: Dict) -> Error:
