+++ legal-api/src/legal_api/services/utils.py
